# Speed up GitHub Actions