--- a/.github/workflows/check.yml
+++ b/.github/workflows/check.yml
@@ -4,7 +4,7 @@ on:
   push:
 
 jobs:
-  format:
+  prettier:
     name: Prettier
     runs-on: ubuntu-20.04
     steps: