--- a/.github/workflows/check.yml
+++ b/.github/workflows/check.yml
@@ -24,6 +24,8 @@ jobs:
         uses: actions/setup-node@v3
         with: {node-version: '${{ matrix.node }}', cache: npm}
       - if: steps.cache-restore.outputs.cache-hit != 'true'
+        run: npm ci
+      - if: steps.cache-restore.outputs.cache-hit != 'true'
         uses: actions/cache/save@v3
         with: 
           path: node_modules/