--- a/.github/workflows/check.yml
+++ b/.github/workflows/check.yml
@@ -37,8 +37,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ ubuntu-22.04, macos-11 ]
-        # I'm a little worried about using allowing the minor version to fluctuate across cache hits...
+        os: [ ubuntu-22.04 ]
         node: [ ^16.0 ]
     steps:
       - uses: actions/checkout@v3.3.0