--- a/detox.config.js
+++ b/detox.config.js
@@ -87,7 +87,8 @@ module.exports = {
 		},
 		'android.emu.debug': {
 			binaryPath: 'android/app/build/outputs/apk/debug/app-debug.apk',
-			build: 'cd android && ./gradlew assembleDebug assembleAndroidTest -DtestBuildType=debug && cd ..',
+			build:
+				'cd android && ./gradlew assembleDebug assembleAndroidTest -DtestBuildType=debug && cd ..',
 			type: 'android.emulator',
 			name: 'emu',
 		},